moving utils to comms interface package
