+++ kernel/packages/shared/comm-interface/utils.ts
